feat(commercial): Phase F networking extraction to fleet modules
Move cross-node networking (models, ports, HTTP bus, knowledge sync,
plasticity, brick usage cache) and AgentBusNetworkBridge from open
src/ into Nexo.Commercial.Fleet.Contracts and Fleet.Infrastructure.

Relocate networking and execution tests to Nexo.Commercial.Tests.Fleet.
Expose AddNexoCommercialFleetNetworking() for commercial fleet hosts.
Update LICENSING.md and CommercialExtractionPlan.md for Phase F.
