+++ commercial/src/Nexo.Commercial.Fleet.Contracts/Networking/Models/AdaptiveBrickCacheStats.cs
