+++ commercial/src/Nexo.Commercial.Fleet.Contracts/Networking/Models/BrickUsageRecord.cs
