+++ commercial/src/Nexo.Commercial.Fleet.Contracts/Networking/Models/BrickUsageStats.cs
