+++ commercial/src/Nexo.Commercial.Fleet.Contracts/Networking/Models/KnowledgeChunk.cs
