+++ commercial/src/Nexo.Commercial.Fleet.Contracts/Networking/Models/KnowledgeSyncStatus.cs
