+++ commercial/src/Nexo.Commercial.Fleet.Contracts/Networking/Models/NetworkAgentEntry.cs
