+++ commercial/src/Nexo.Commercial.Fleet.Contracts/Networking/Models/NetworkEvent.cs
