+++ commercial/src/Nexo.Commercial.Fleet.Contracts/Networking/Models/NetworkEventTypes.cs
