+++ commercial/src/Nexo.Commercial.Fleet.Contracts/Networking/Models/PlasticityMetrics.cs
