+++ commercial/src/Nexo.Commercial.Fleet.Contracts/Networking/Ports/IAdaptiveBrickCache.cs
