+++ commercial/src/Nexo.Commercial.Fleet.Contracts/Networking/Ports/IKnowledgeChunkStore.cs
