+++ commercial/src/Nexo.Commercial.Fleet.Contracts/Nexo.Commercial.Fleet.Contracts.csproj
@@ -10,5 +10,6 @@
 
   <ItemGroup>
     <ProjectReference Include="../../../src/Nexo.Core.Application/Nexo.Core.Application.csproj" />
+    <ProjectReference Include="../../../src/Nexo.Core.Domain/Nexo.Core.Domain.csproj" />
   </ItemGroup>
 </Project>
