+++ commercial/src/Nexo.Commercial.Fleet.Infrastructure/Execution/AdaptiveBrickCache.cs
