+++ commercial/src/Nexo.Commercial.Fleet.Infrastructure/Execution/BrickUsageTracker.cs
