+++ commercial/src/Nexo.Commercial.Fleet.Infrastructure/Networking/HttpNetworkAgentDirectory.cs
