+++ commercial/src/Nexo.Commercial.Fleet.Infrastructure/Networking/InMemoryKnowledgeChunkStore.cs
