+++ commercial/src/Nexo.Commercial.Fleet.Infrastructure/Networking/NetworkAgentDirectoryOptions.cs
