+++ commercial/src/Nexo.Commercial.Fleet.Infrastructure/Networking/NetworkNegotiationService.cs
