+++ commercial/src/Nexo.Commercial.Fleet.Infrastructure/Nexo.Commercial.Fleet.Infrastructure.csproj
@@ -15,6 +15,8 @@
   <ItemGroup>
     <ProjectReference Include="../Nexo.Commercial.Fleet.Contracts/Nexo.Commercial.Fleet.Contracts.csproj" />
     <ProjectReference Include="../../../src/Nexo.Core.Application/Nexo.Core.Application.csproj" />
+    <ProjectReference Include="../../../src/Nexo.Core.Domain/Nexo.Core.Domain.csproj" />
+    <ProjectReference Include="../../../src/Nexo.Orchestration/Nexo.Orchestration.csproj" />
   </ItemGroup>
 
   <ItemGroup>
